refactor: 版本迁移统一入口——src/migrate.js 注册表式调度器（archive-old-logs / config-schema / junction-converge）
把散落的「版本迁移 / 历史遗留收敛」动作收敛进唯一入口 src/migrate.js（有序注册表 +
runMigrations(cfg, opts) 调度器，每步幂等、失败不阻断后续；新增迁移只加一条注册表项）：

- archive-old-logs：旧日志归档压缩（latest.log 残留归档 + 时间戳 .log → .log.zst，全部
  保留不删除），原 index.js onload 内联 → 改经 g.runMigrations 单例调用（index.js 不静态
  import migrate.js，遵守分发纪律；须在建新会话文件之前执行）
- config-schema：config.json schema 初始化/升级（不存在时按 manifest 默认值生成，幂等不
  覆盖），原 lifecycle.js ensureConfigJson → 迁入 migrate.js，startWebHostFromPlugin 与
  dsh-run.js doExecute 改调 runMigrations
- junction-converge：cordis junction 旧名清理 + @dsh-hanako scope 无条件收敛，原
  lifecycle.js ensureWebHost 内闭包 → 迁入 migrate.js，ensureWebHost 每次 spawn 前改调
  runMigrations

未纳入（保持原地，取舍见 migrate.js 头注释）：update-result.json（运行期状态文件）、
newWebLogPath 兜底、provider 路由/bus 等运行期能力。

分发：migrate.js 被 lifecycle.js / dsh-run.js 静态 import，随单 bundle 内联进
dist/index.js；g.runMigrations 顶层挂 globalThis 单例（与 g.startWebHost 同纪律）。
行为保持：日志归档先于建会话文件、config.json 幂等不覆盖、junction 无条件收敛均不变。
验证：node scripts/build.mjs 通过（dist 不进提交）。
